Trait hycon_rust::consensus::BlockForkChoice

pub trait BlockForkChoice {
    fn fork_choice(&self, other: &Self) -> Ordering;
}

BlockForkChoice Trait

Acts as a comparator between two blocks to give precedence to one over the other in the event of a forking event

Required Methods

Fork Choice

fn fork_choice(&self, other: &Self) -> std::cmp::Ordering

Arguments

other - some other block to compare to

Returns

An ordering between self and other

Required methods

fn fork_choice(&self, other: &Self) -> Ordering

Loading content...